Move tests of ElementConsistencyRule to more appropriate module
The ElementConsistencyRule is not exactly a unique name validation,
and the other existing test of it was broken and marked for skipping,
so moving this makes a fair bit of sense.
